+++ ccore/src/container/adjacency_connector.hpp
@@ -240,7 +240,7 @@ public:
 			    m_connector(index, upper_index, output_adjacency_collection);
 		    }
 
-		    if (lower_index < output_adjacency_collection.size()) {
+		    if (lower_index < (int) output_adjacency_collection.size()) {
 			    m_connector(index, lower_index, output_adjacency_collection);
 		    }
 
@@ -248,7 +248,7 @@ public:
 			    m_connector(index, left_index, output_adjacency_collection);
 		    }
 
-            if ((right_index < output_adjacency_collection.size()) && (std::ceil(right_index / width) == node_row_index)) {
+            if ((right_index < (int) output_adjacency_collection.size()) && (std::ceil(right_index / width) == node_row_index)) {
 			    m_connector(index, right_index, output_adjacency_collection);
 		    }
 	    }
@@ -288,14 +288,14 @@ public:
     virtual void create_grid_eight_connections(const size_t width, const size_t height, TypeCollection & output_adjacency_collection) {
 	    create_grid_four_connections(width, height, output_adjacency_collection);	/* create connection with right, upper, left, lower neighbor */
 
-	    for (int index = 0; index < output_adjacency_collection.size(); index++) {
+	    for (int index = 0; index < (int) output_adjacency_collection.size(); index++) {
             const int upper_left_index = index - width - 1;
             const int upper_right_index = index - width + 1;
             
             const int lower_left_index = index + width - 1;
             const int lower_right_index = index + width + 1;
             
-            const int node_row_index = std::floor(index / width);
+            const int node_row_index = (int) std::floor(index / width);
             const int upper_row_index = node_row_index - 1;
             const int lower_row_index = node_row_index + 1;
 
@@ -307,11 +307,11 @@ public:
 			    m_connector(index, upper_right_index, output_adjacency_collection);
 		    }
 
-            if ((lower_left_index < output_adjacency_collection.size()) && (std::floor(lower_left_index / width) == lower_row_index)) {
+            if ((lower_left_index < (int) output_adjacency_collection.size()) && (std::floor(lower_left_index / width) == lower_row_index)) {
 			    m_connector(index, lower_left_index, output_adjacency_collection);
 		    }
 
-            if ((lower_right_index < output_adjacency_collection.size()) && (std::floor(lower_right_index / width) == lower_row_index)) {
+            if ((lower_right_index < (int) output_adjacency_collection.size()) && (std::floor(lower_right_index / width) == lower_row_index)) {
 			    m_connector(index, lower_right_index, output_adjacency_collection);
 		    }
 	    }
